+++ kernel/inc/task.h
@@ -71,15 +71,15 @@ void init_jiffies(void);
 void init_task(void);
 void create_all_task(void);
 void forkyi(struct task_struct *pbt, struct task_struct *pt);
-struct task_struct *do_forkyi(char *name, task_entry fn, int idx, unsigned long *ppd);
+struct task_struct *do_forkyi(char *name, task_entry fn, int idx, unsigned int *ppd);
 void switch_context(struct task_struct *prev, struct task_struct *next);
 void schedule(void *arg);
 void dequeue_se_to_exit(struct cfs_rq *rq, struct sched_entity *se);
 void enqueue_se_to_runq(struct cfs_rq *rq, struct sched_entity *se, bool update);
 void dequeue_se_to_waitq(struct cfs_rq *rq, struct sched_entity *se, bool update);
 void drop_usrtask();
-void init_idletask(unsigned long *ppd);
-void init_shell(unsigned long *ppd);
+void init_idletask(unsigned int *ppd);
+void init_shell(unsigned int *ppd);
 void update_se();
 void set_priority(struct task_struct *pt, uint32_t pri);
